+++ modules/apps/foundation/apio-architect/apio-architect-api/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Activator: com.liferay.apio.architect.internal.WebsphereApiSupport
+Bundle-Activator: com.liferay.apio.architect.websphere.WebsphereApiSupport
 Bundle-Name: Liferay Apio Architect API
 Bundle-SymbolicName: com.liferay.apio.architect.api
 Bundle-Version: 1.0.0
